+++ kleides_mfa/registry.py
@@ -5,7 +5,7 @@ from collections import namedtuple
 
 from django.forms import modelform_factory
 from django.utils.text import slugify
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from .forms import DeviceUpdateForm
 from .settings import PLUGIN_PRIORITY